requirements.txt
flask
flask-cors
requests
python-dotenv
llama-index